feat: ranking add / update / getNow 가능

+++ BE/relpl/src/main/java/com/ssafy/relpl/dto/response/RankingEntry.java
@@ -10,5 +10,5 @@ import lombok.*;
 @ToString
 public class RankingEntry {
         private String nickname;
-        private int distance;
+        private double distance;
 }
